core-to-plc: big reorganization
This reorganizes core-to-plc so there are more, smaller files. I've
wanted to do this for a while, but unfortunately there are some big
recursive cycles. I've just bitten the bullet and added `.hs-boot` files
for these.
