+++ .hlint.yaml
@@ -18,7 +18,7 @@
 - error: {lhs: "maybe mempty", rhs: "foldMap", name: "Use foldMap"}
 - error: {lhs: "mconcat", rhs: "fold", name: "Generalize mconcat"}
 
-- ignore: {name: Reduce duplication, within: [Language.PlutusCore.Renamer, Language.PlutusCore.Constant.Prelude, Language.PlutusCore.StdLib.Data.Bool, Language.PlutusCore.StdLib.Data.ChurchNat, Language.PlutusCore.StdLib.Data.Function, Language.PlutusCore.StdLib.Data.List, Language.PlutusCore.StdLib.Data.Nat, Language.PlutusCore.Pretty.Readable, Language.Plutus.CoreToPLC, Evaluation.CkMachine, Spec.Crowdfunding, Spec.Vesting, Language.Plutus.Lift]}
+- ignore: {name: Reduce duplication, within: [Language.PlutusCore.Renamer, Language.PlutusCore.Constant.Prelude, Language.PlutusCore.StdLib.Data.Bool, Language.PlutusCore.StdLib.Data.ChurchNat, Language.PlutusCore.StdLib.Data.Function, Language.PlutusCore.StdLib.Data.List, Language.PlutusCore.StdLib.Data.Nat, Language.PlutusCore.Pretty.Readable, Language.Plutus.CoreToPLC.Compiler.Binders, Language.Plutus.CoreToPLC.Compiler.Type, Evaluation.CkMachine, Spec.Crowdfunding, Spec.Vesting, Language.Plutus.Lift]}
 - ignore: {name: Redundant $, within: [Evaluation.Constant.Success, Language.PlutusCore.Generators.Internal.TypedBuiltinGen]}
 - ignore: {name: Redundant bracket, within: [Language.Plutus.TH]}
 # this is rarely an improvement, also ignored in cardano
