+++ core-to-plc/src/Language/Plutus/CoreToPLC/Compiler/Type.hs
@@ -2,159 +2,54 @@
 {-# LANGUAGE FlexibleContexts  #-}
 {-# LANGUAGE LambdaCase        #-}
 {-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE TemplateHaskell   #-}
 {-# LANGUAGE ViewPatterns      #-}
 
-module Language.Plutus.CoreToPLC where
+-- | Functions for compiling GHC types into PlutusCore types, as well as compiling constructors,
+-- matchers, and pattern match alternatives.
+module Language.Plutus.CoreToPLC.Compiler.Type (
+    convType,
+    getDataCons,
+    getConstructors,
+    getConstructorsInstantiated,
+    getMatch,
+    getMatchInstantiated,
+    convAlt,
+    NewtypeCoercion (..),
+    splitNewtypeCoercion) where
 
-import           Language.Plutus.CoreToPLC.Builtins
-import           Language.Plutus.CoreToPLC.Definitions
+import           Language.Plutus.CoreToPLC.Compiler.Binders
+import           Language.Plutus.CoreToPLC.Compiler.Builtins
+import           Language.Plutus.CoreToPLC.Compiler.Definitions
+import {-# SOURCE #-} Language.Plutus.CoreToPLC.Compiler.Expr
+import           Language.Plutus.CoreToPLC.Compiler.Kind
+import           Language.Plutus.CoreToPLC.Compiler.Names
+import           Language.Plutus.CoreToPLC.Compiler.Types
+import           Language.Plutus.CoreToPLC.Compiler.Utils
 import           Language.Plutus.CoreToPLC.Error
 import           Language.Plutus.CoreToPLC.Laziness
 import           Language.Plutus.CoreToPLC.PLCTypes
-import qualified Language.Plutus.CoreToPLC.Primitives     as Prims
-import           Language.Plutus.CoreToPLC.Types
+import           Language.Plutus.CoreToPLC.Utils
 
-import qualified CoreUtils                                as GHC
-import qualified GhcPlugins                               as GHC
-import qualified Kind                                     as GHC
-import qualified MkId                                     as GHC
-import qualified Pair                                     as GHC
-import qualified PrelNames                                as GHC
-import qualified PrimOp                                   as GHC
-import qualified TysPrim                                  as GHC
+import qualified GhcPlugins                                     as GHC
+import qualified Pair                                           as GHC
+import qualified PrelNames                                      as GHC
+import qualified TysPrim                                        as GHC
 
-import qualified Language.PlutusCore                      as PLC
+import qualified Language.PlutusCore                            as PLC
 import           Language.PlutusCore.MkPlc
 import           Language.PlutusCore.Quote
-import qualified Language.PlutusCore.StdLib.Data.Function as Function
 
-import qualified Language.Haskell.TH.Syntax               as TH
-
-import           Control.Monad.Except
 import           Control.Monad.Reader
 import           Control.Monad.State
 
-import qualified Data.ByteString.Lazy                     as BSL
-import           Data.Char
 import           Data.Foldable
-import qualified Data.List.NonEmpty                       as NE
-import qualified Data.Map                                 as Map
-import qualified Data.Text                                as T
-import qualified Data.Text.Encoding                       as TE
+import qualified Data.List.NonEmpty                             as NE
+import qualified Data.Map                                       as Map
 import           Lens.Micro
 
-import           Data.List                                (elemIndex, reverse, sortBy)
+import           Data.List                                      (elemIndex, reverse, sortBy)
 
-{- Note [System FC and System FW]
-Haskell uses system FC, which includes type equalities and coercions.
-
-PLC does *not* have coercions in particular. However, PLC also does not have a nominal
-type system - everything is constructed via operators on base types, so we have no
-need for coercions to convert between newtypes and datatypes.
-
-So we mostly ignore coercions. The one place that I know of where the mismatch hurts
-us is that GHC uses the `Any` type (coercible to and from anything) for unconstrained
-function variables, e.g. in polymorphic lambdas. This is annoying for us, since we
-really want the version with explicit type abstraction. I don't currently have a fix
-for this.
--}
-
-{- Note [Iterated abstraction and application]
-PLC doesn't expose iterated abstraction and application.
-
-We typically build these with a fold.
-- Iterated application uses a *left* fold, since we want to apply the first variable
-first.
-- Iterated abstraction uses a *right* fold, since we want to bind the first
-variable *last* (so it is on the outside, so will be first when applying).
--}
-
-strToBs :: String -> BSL.ByteString
-strToBs = BSL.fromStrict . TE.encodeUtf8 . T.pack
-
-bsToStr :: BSL.ByteString -> String
-bsToStr = T.unpack . TE.decodeUtf8 . BSL.toStrict
-
-sdToTxt :: (MonadReader ConvertingContext m) => GHC.SDoc -> m T.Text
-sdToTxt sd = do
-  ConvertingContext { ccFlags=flags } <- ask
-  pure $ T.pack $ GHC.showSDoc flags sd
-
-throwSd :: (MonadError ConvError m, MonadReader ConvertingContext m) => (T.Text -> Error ()) -> GHC.SDoc -> m a
-throwSd constr = (throwPlain . constr) <=< sdToTxt
-
--- Names and scopes
-
-lookupName :: Scope -> GHC.Name -> Maybe (PLC.Name ())
-lookupName (Scope ns _) n = Map.lookup n ns
-
-{- Note [PLC names]
-We convert names from Haskell names quite frequently here, but PLC admits a much
-smaller set of valid identifiers. We compromise by mangling the identifier, but
-in the long run it would be nice to have a more principled encoding so we can
-support unicode identifiers as well.
--}
-
-safeFreshName :: MonadQuote m => String -> m (PLC.Name ())
-safeFreshName s
-    -- some special cases
-    | s == ":" = safeFreshName "cons"
-    | s == "[]" = safeFreshName "list"
-    | otherwise =
-          let
-              -- See Note [PLC names]
-              -- first strip out disallowed characters
-              stripped = filter (\c -> isLetter c || isDigit c || c == '_' || c == '`') s
-              -- now fix up some other bits
-              fixed = case stripped of
-                -- empty name, just put something to mark that
-                []      -> "bad_name"
-                -- can't start with these
-                ('`':_) -> "p" ++ stripped
-                ('_':_) -> "p" ++ stripped
-                n       -> n
-          in liftQuote $ freshName () $ strToBs fixed
-
-convNameFresh :: MonadQuote m => GHC.Name -> m (PLC.Name ())
-convNameFresh n = safeFreshName $ GHC.getOccString n
-
-convVarFresh :: Converting m => GHC.Var -> m (PLCType, PLC.Name ())
-convVarFresh v = do
-    t' <- convType $ GHC.varType v
-    n' <- convNameFresh $ GHC.getName v
-    pure (t', n')
-
-lookupTyName :: Scope -> GHC.Name -> Maybe (PLC.TyName ())
-lookupTyName (Scope _ tyns) n = Map.lookup n tyns
-
-safeFreshTyName :: MonadQuote m => String -> m (PLC.TyName ())
-safeFreshTyName s = PLC.TyName <$> safeFreshName s
-
-convTyNameFresh :: MonadQuote m => GHC.Name -> m (PLC.TyName ())
-convTyNameFresh n = PLC.TyName <$> convNameFresh n
-
-convTyVarFresh :: Converting m => GHC.TyVar -> m (PLC.TyName (), PLC.Kind ())
-convTyVarFresh v = do
-    k' <- convKind $ GHC.tyVarKind v
-    t' <- convTyNameFresh $ GHC.getName v
-    pure (t', k')
-
-pushName :: GHC.Name -> PLC.Name () -> ScopeStack -> ScopeStack
-pushName ghcName n stack = let Scope ns tyns = NE.head stack in Scope (Map.insert ghcName n ns) tyns NE.<| stack
-
-pushTyName :: GHC.Name -> PLC.TyName () -> ScopeStack -> ScopeStack
-pushTyName ghcName n stack = let Scope ns tyns = NE.head stack in Scope ns (Map.insert ghcName n tyns) NE.<| stack
-
--- Types and kinds
-
-convKind :: Converting m => GHC.Kind -> m (PLC.Kind ())
-convKind k = withContextM (sdToTxt $ "Converting kind:" GHC.<+> GHC.ppr k) $ case k of
-    -- this is a bit weird because GHC uses 'Type' to represent kinds, so '* -> *' is a 'TyFun'
-    (GHC.isStarKind -> True)              -> pure $ PLC.Type ()
-    (GHC.splitFunTy_maybe -> Just (i, o)) -> PLC.KindArrow () <$> convKind i <*> convKind o
-    _                                     -> throwSd UnsupportedError $ "Kind:" GHC.<+> GHC.ppr k
-
+-- Types
 
 convType :: Converting m => GHC.Type -> m PLCType
 convType t = withContextM (sdToTxt $ "Converting type:" GHC.<+> GHC.ppr t) $ do
@@ -289,12 +184,6 @@ isSelfRecursiveTyCon tc = do
     -- See Note [Detecting recursive types in GHC]
     usedTcs <- getUsedTcs tc
     pure $ tc `elem` usedTcs
-
-isSelfRecursiveType :: (Converting m) => GHC.Type -> m Bool
-isSelfRecursiveType t = case t of
-    -- the only way it can be *self* recursive is if it's type constructor
-    (GHC.splitTyConApp_maybe -> Just (tc, _)) -> isSelfRecursiveTyCon tc
-    _                                         -> pure False
 
 -- | Split out a recursive type (constructed by us) into its pattern functor and the name of the variable.
 -- This is a bit of a hack, since it relies on us knowing that the structure will be some number of applications
@@ -747,429 +636,3 @@ convAlt mustDelay dc (alt, vars, body) = case alt of
     -- See Note [Iterated abstraction and application]
     -- See Note [Case expressions and laziness]
     GHC.DataAlt _ -> mkIterLamAbsScoped vars (convExpr body >>= maybeDelay mustDelay)
-
--- Literals and primitives
-
-{- Note [Literals]
-GHC's literals and primitives are a bit of a pain, since they not only have a Literal
-containing the actual data, but are wrapped in special functions (often ending in the magic #).
-
-This is a pain to recognize.
--}
-
-convLiteral :: Converting m => GHC.Literal -> m (PLC.Constant ())
-convLiteral l = case l of
-    -- TODO: better sizes
-    GHC.MachInt64 i    -> pure $ PLC.BuiltinInt () haskellIntSize i
-    GHC.MachInt i      -> pure $ PLC.BuiltinInt () haskellIntSize i
-    GHC.MachStr bs     -> pure $ PLC.BuiltinBS () haskellBSSize (BSL.fromStrict bs)
-    GHC.LitInteger _ _ -> throwPlain $ UnsupportedError "Literal (unbounded) integer"
-    GHC.MachWord _     -> throwPlain $ UnsupportedError "Literal word"
-    GHC.MachWord64 _   -> throwPlain $ UnsupportedError "Literal word64"
-    GHC.MachChar _     -> throwPlain $ UnsupportedError "Literal char"
-    GHC.MachFloat _    -> throwPlain $ UnsupportedError "Literal float"
-    GHC.MachDouble _   -> throwPlain $ UnsupportedError "Literal double"
-    GHC.MachLabel {}   -> throwPlain $ UnsupportedError "Literal label"
-    GHC.MachNullAddr   -> throwPlain $ UnsupportedError "Literal null"
-
-isPrimitiveWrapper :: GHC.Id -> Bool
-isPrimitiveWrapper i = case GHC.idDetails i of
-    GHC.DataConWorkId dc -> isPrimitiveDataCon dc
-    GHC.VanillaId        -> GHC.getName i == GHC.unpackCStringName
-    _                    -> False
-
-isPrimitiveDataCon :: GHC.DataCon -> Bool
-isPrimitiveDataCon dc = dc == GHC.intDataCon
-
--- These never seem to come up, rather we get the typeclass operations. Not sure if we need them.
-convPrimitiveOp :: (Converting m) => GHC.PrimOp -> m PLCTerm
-convPrimitiveOp = \case
-    GHC.IntAddOp  -> pure $ mkIntFun PLC.AddInteger
-    GHC.IntSubOp  -> pure $ mkIntFun PLC.SubtractInteger
-    GHC.IntMulOp  -> pure $ mkIntFun PLC.MultiplyInteger
-    -- check this one
-    GHC.IntQuotOp -> pure $ mkIntFun PLC.DivideInteger
-    GHC.IntRemOp  -> pure $ mkIntFun PLC.RemainderInteger
-    GHC.IntGtOp   -> pure $ mkIntRel PLC.GreaterThanInteger
-    GHC.IntGeOp   -> pure $ mkIntRel PLC.GreaterThanEqInteger
-    GHC.IntLtOp   -> pure $ mkIntRel PLC.LessThanInteger
-    GHC.IntLeOp   -> pure $ mkIntRel PLC.LessThanEqInteger
-    GHC.IntEqOp   -> pure $ mkIntRel PLC.EqInteger
-    po            -> throwSd UnsupportedError $ "Primitive operation:" GHC.<+> GHC.ppr po
-
--- Typeclasses
-
-convEqMethod :: (Converting m) => GHC.Name -> m PLCTerm
-convEqMethod name
-    | name == GHC.eqName = pure $ mkIntRel PLC.EqInteger
-    | otherwise = throwSd UnsupportedError $ "Eq method:" GHC.<+> GHC.ppr name
-
-convOrdMethod :: (Converting m) => GHC.Name -> m PLCTerm
-convOrdMethod name
-    -- only this one has a name defined in the lib??
-    | name == GHC.geName = pure $ mkIntRel PLC.GreaterThanEqInteger
-    | GHC.getOccString name == ">" = pure $ mkIntRel PLC.GreaterThanInteger
-    | GHC.getOccString name == "<=" = pure $ mkIntRel PLC.LessThanEqInteger
-    | GHC.getOccString name == "<" = pure $ mkIntRel PLC.LessThanInteger
-    | otherwise = throwSd UnsupportedError $ "Ord method:" GHC.<+> GHC.ppr name
-
-convNumMethod :: (Converting m) => GHC.Name -> m PLCTerm
-convNumMethod name
-    -- only this one has a name defined in the lib??
-    | name == GHC.minusName = pure $ mkIntFun PLC.SubtractInteger
-    | GHC.getOccString name == "+" = pure $ mkIntFun PLC.AddInteger
-    | GHC.getOccString name == "*" = pure $ mkIntFun PLC.MultiplyInteger
-    | otherwise = throwSd UnsupportedError $ "Num method:" GHC.<+> GHC.ppr name
-
--- Plutus primitives
-
-{- Note [Mapping primitives]
-We want the user to be able to call the Plutus primitives as normal Haskell functions.
-
-To do this, we provide a library of such functions, and then make a map from their TH names (which we can
-derive from the real function declarations, to be sure they match up), to the implementations. We then
-need to do some work in the GHC Core monad to translate those mappings into mappings from Core names.
--}
-
-primitiveTermAssociations :: [(TH.Name, Quote (PLC.Term PLC.TyName PLC.Name ()))]
-primitiveTermAssociations = [
-    ('Prims.concatenate, pure $ instSize haskellIntSize $ mkConstant PLC.Concatenate)
-    , ('Prims.takeByteString, pure $ instSize haskellBSSize $ instSize haskellIntSize $ mkConstant PLC.TakeByteString)
-    , ('Prims.dropByteString, pure $ instSize haskellBSSize $ instSize haskellIntSize $ mkConstant PLC.DropByteString)
-    , ('Prims.sha2_256, pure $ instSize haskellBSSize $ mkConstant PLC.SHA2)
-    , ('Prims.sha3_256, pure $ instSize haskellBSSize $ mkConstant PLC.SHA3)
-    , ('Prims.verifySignature, pure $ instSize haskellBSSize $ instSize haskellBSSize $ instSize haskellBSSize $ mkConstant PLC.VerifySignature)
-    , ('Prims.equalsByteString, pure $ mkBsRel PLC.EqByteString)
-    , ('Prims.txhash, pure $ mkConstant PLC.TxHash)
-    , ('Prims.blocknum, pure $ instSize haskellIntSize $ mkConstant PLC.BlockNum)
-    -- we're representing error at the haskell level as a polymorphic function, so do the same here
-    , ('Prims.error, errorFunc)
-    ]
-
-primitiveTypeAssociations :: [(TH.Name, Quote (PLC.Type PLC.TyName ()))]
-primitiveTypeAssociations = [
-    (''Prims.ByteString, pure $ appSize haskellBSSize $ PLC.TyBuiltin () PLC.TyByteString)
-    ]
-
--- | The function 'error :: forall a . () -> a'.
-errorFunc :: Quote (PLC.Term PLC.TyName PLC.Name ())
-errorFunc = do
-    n <- freshTyName () "e"
-    -- see Note [Value restriction]
-    mangleTyAbs $ PLC.TyAbs () n (PLC.Type ()) (PLC.Error () (PLC.TyVar () n))
-
--- | The type 'forall a. () -> a'.
-errorTy :: Quote (PLC.Type PLC.TyName ())
-errorTy = do
-    tyname <- safeFreshTyName "a"
-    mangleTyForall $ PLC.TyForall () tyname (PLC.Type ()) (PLC.TyVar () tyname)
-
--- Value restriction
-
-{- Note [Value restriction]
-Plutus Core has the traditional *value restriction* on type abstractions - namely, the
-body of a type abstraction must be a value.
-
-This causes problems for us because *Haskell* has no such thing. (There is the monomorphism
-restriciton, which is similar, but not as restrictive, so it doesn't help us.)
-
-There are two approaches to solving this problem. Currently we use "passing on the value restriction",
-but I include a description of mangling to illustrate why the current solution is preferable.
-
--- Mangling
-
-We can get around this by delaying the body of all our abstractions, thus making them lambdas, which are values.
-This then means that we need to change the *types* of all foralls to include the unit argument,
-and all instantiations to force the value.
-
-We need to do this everywhere so that the translation of the users' program remains
-well-typed. Consider
-
-runST :: (forall s. ST s a) -> a
-
-myCalc :: Int
-myCalc = runST $ pure 1
-
-Converting `pure 1`, we would normally turn it into something of type
-
-(all s (type) [ST s Int])
-
-After mangling, this becomes
-
-(all s (type) (fun unit [ST s Int]))
-
-This means we had better convert `runST` into something that expects things of that type instead of
-the original type! And we need to add forces to the instantiations inside `runST` otherwise its
-body won't be well-typed too.
-
-Note that it's no good to convert some other abstraction where the body is already a value without the delay -
-if we did that, then `runST` would be in an impossible place, since it would need to take
-either mangled or non-mangled abstractions. The only way we can get away with this without doing complicated
-accounting is to do it uniformly: i.e. absolutely everywhere whether we need it in that
-case or not.
-
-We need to do this even for the abstractions in our generated matchers for Scott-encoding,
-because constructors are user-visible and so can be passed to functions, which might expect them to be
-mangled.
-
--- Passing on the value restriction
-
-An alternative approach would be to pass on the value restriction to the client Haskell code.
-This has the advantage of much simpler codegen. However, it has a few annoying cases:
-
-- We can't provide `error :: forall a. a` any more.
-- We can't handle `void# :: forall a. a` well any more.
-
-The first is easy: we just make the primitive be mangled `error :: forall a. () -> a`.
-
-The second is a pain. GHC *does* use the polymorphic void. So we take the ad-hoc
-expedient of mangling *just* the 'Void#' type. Since nothing can use this "for real",
-there are no uses to go wrong so long as we change the type uniformly. This is a
-bit of a hack, though.
--}
-
--- See Note [Value restriction]
-mangleTyForall :: (MonadQuote m) => PLCType -> m PLCType
-mangleTyForall = \case
-    PLC.TyForall a t k body -> PLC.TyForall a t k <$> delayType body
-    x -> pure x
-
--- See Note [Value restriction]
-mangleTyAbs :: (MonadQuote m) => PLCTerm -> m PLCTerm
-mangleTyAbs = \case
-    PLC.TyAbs a t k body -> PLC.TyAbs a t k <$> delay body
-    x -> pure x
-
--- Binder helpers
-
--- | Builds a lambda, binding the given variable to a name that
--- will be in scope when running the second argument.
-mkLamAbsScoped :: Converting m => GHC.Var -> m PLCTerm -> m PLCTerm
-mkLamAbsScoped v body = do
-    let ghcName = GHC.getName v
-    (t', n') <- convVarFresh v
-    body' <- local (\c -> c {ccScopes=pushName ghcName n' (ccScopes c)}) body
-    pure $ PLC.LamAbs () n' t' body'
-
-mkIterLamAbsScoped :: Converting m => [GHC.Var] -> m PLCTerm -> m PLCTerm
-mkIterLamAbsScoped vars body = foldr (\v acc -> mkLamAbsScoped v acc) body vars
-
--- | Builds a type abstraction, binding the given variable to a name that
--- will be in scope when running the second argument.
-mkTyAbsScoped :: Converting m => GHC.Var -> m PLCTerm -> m PLCTerm
-mkTyAbsScoped v body = do
-    let ghcName = GHC.getName v
-    (t', k') <- convTyVarFresh v
-    body' <- local (\c -> c {ccScopes=pushTyName ghcName t' (ccScopes c)}) body
-    ConvertingContext {ccOpts=opts} <- ask
-    -- we sometimes need to turn this off, as checking for term values also checks for normalized types at the moment
-    unless (not (coCheckValueRestriction opts) || PLC.isTermValue body') $ throwPlain $ ValueRestrictionError "Type abstraction body is not a value"
-    pure $ PLC.TyAbs () t' k' body'
-
-mkIterTyAbsScoped :: Converting m => [GHC.Var] -> m PLCTerm -> m PLCTerm
-mkIterTyAbsScoped vars body = foldr (\v acc -> mkTyAbsScoped v acc) body vars
-
--- | Builds a forall, binding the given variable to a name that
--- will be in scope when running the second argument.
-mkTyForallScoped :: Converting m => GHC.Var -> m PLCType -> m PLCType
-mkTyForallScoped v body = do
-    let ghcName = GHC.getName v
-    (t', k') <- convTyVarFresh v
-    body' <- local (\c -> c {ccScopes=pushTyName ghcName t' (ccScopes c)}) body
-    pure $ PLC.TyForall () t' k' body'
-
-mkIterTyForallScoped :: Converting m => [GHC.Var] -> m PLCType -> m PLCType
-mkIterTyForallScoped vars body = foldr (\v acc -> mkTyForallScoped v acc) body vars
-
--- | Builds a type lambda, binding the given variable to a name that
--- will be in scope when running the second argument.
-mkTyLamScoped :: Converting m => GHC.Var -> m PLCType -> m PLCType
-mkTyLamScoped v body = do
-    let ghcName = GHC.getName v
-    (t', k') <- convTyVarFresh v
-    body' <- local (\c -> c {ccScopes=pushTyName ghcName t' (ccScopes c)}) body
-    pure $ PLC.TyLam () t' k' body'
-
-mkIterTyLamScoped :: Converting m => [GHC.Var] -> m PLCType -> m PLCType
-mkIterTyLamScoped vars body = foldr (\v acc -> mkTyLamScoped v acc) body vars
-
-{- Note [Recursive lets]
-We need to define these with a fixpoint. We can derive a fixpoint operator for values
-already.
-
-However, we also need to work out how to encode recursion over multiple values simultaneously.
-The answer is simple - we pass them through as a tuple.
-
-Overall, the translation looks like this. We convert this:
-
-let rec
-  f_1 = b_1
-  ...
-  f_i = b_i
-in
-  result
-
-into this:
-
-($fixN i$ (\choose f1 ... fi . choose b_1 ... b_i))
-(\f1 ... fi . result)
--}
-
--- Expressions
-
-convExpr :: Converting m => GHC.CoreExpr -> m PLCTerm
-convExpr e = withContextM (sdToTxt $ "Converting expr:" GHC.<+> GHC.ppr e) $ do
-    -- See Note [Scopes]
-    ConvertingContext {ccPrimTerms=prims, ccScopes=stack} <- ask
-    let top = NE.head stack
-    case e of
-        -- See Note [Literals]
-        GHC.App (GHC.Var (isPrimitiveWrapper -> True)) arg -> convExpr arg
-        -- special typeclass method calls
-        GHC.App (GHC.App
-                -- eq class method
-                (GHC.Var n@(GHC.idDetails -> GHC.ClassOpId ((==) GHC.eqClassName . GHC.getName -> True)))
-                -- we only support applying to int
-                (GHC.Type (GHC.eqType GHC.intTy -> True)))
-            -- last arg is typeclass dictionary
-            _ -> convEqMethod (GHC.getName n)
-        GHC.App (GHC.App
-                -- ord class method
-                (GHC.Var n@(GHC.idDetails -> GHC.ClassOpId ((==) GHC.ordClassName . GHC.getName -> True)))
-                -- we only support applying to int
-                (GHC.Type (GHC.eqType GHC.intTy -> True)))
-            -- last arg is typeclass dictionary
-            _ -> convOrdMethod (GHC.getName n)
-        GHC.App (GHC.App
-                -- num class method
-                (GHC.Var n@(GHC.idDetails -> GHC.ClassOpId ((==) GHC.numClassName . GHC.getName -> True)))
-                -- we only support applying to int
-                (GHC.Type (GHC.eqType GHC.intTy -> True)))
-            -- last arg is typeclass dictionary
-            _ -> convNumMethod (GHC.getName n)
-        -- void# - values of type void get represented as error, since they should be unreachable
-        GHC.Var n | n == GHC.voidPrimId || n == GHC.voidArgId -> liftQuote errorFunc
-        -- locally bound vars
-        GHC.Var (lookupName top . GHC.getName -> Just name) -> pure $ PLC.Var () name
-        -- Special kinds of id
-        GHC.Var (GHC.idDetails -> GHC.PrimOpId po) -> convPrimitiveOp po
-        GHC.Var (GHC.idDetails -> GHC.DataConWorkId dc) ->
-            let
-                tc = GHC.dataConTyCon dc
-            in do
-                dcs <- getDataCons tc
-                constrs <- getConstructors tc
-
-                -- TODO: this is inelegant
-                index <- case elemIndex dc dcs of
-                    Just i  -> pure i
-                    Nothing -> throwPlain $ ConversionError "Data constructor not in the type constructor's list of constructors!"
-
-                pure $ constrs !! index
-        GHC.Var (flip Map.lookup prims . GHC.getName -> Just term) -> liftQuote term
-        -- the term we get must be closed - we don't resolve most references
-        -- TODO: possibly relax this?
-        GHC.Var n@(GHC.idDetails -> GHC.VanillaId) -> throwSd FreeVariableError $ "Variable:" GHC.<+> GHC.ppr n GHC.$+$ (GHC.ppr $ GHC.idDetails n)
-        GHC.Var n -> throwSd UnsupportedError $ "Variable" GHC.<+> GHC.ppr n GHC.$+$ (GHC.ppr $ GHC.idDetails n)
-        GHC.Lit lit -> PLC.Constant () <$> convLiteral lit
-        -- arg can be a type here, in which case it's a type instantiation
-        GHC.App l (GHC.Type t) -> PLC.TyInst () <$> convExpr l <*> convType t
-        -- otherwise it's a normal application
-        GHC.App l arg -> PLC.Apply () <$> convExpr l <*> convExpr arg
-        -- if we're biding a type variable it's a type abstraction
-        GHC.Lam b@(GHC.isTyVar -> True) body -> mkTyAbsScoped b $ convExpr body
-        -- othewise it's a normal lambda
-        GHC.Lam b body -> mkLamAbsScoped b $ convExpr body
-        GHC.Let (GHC.NonRec b arg) body -> do
-            -- convert it as a lambda
-            a' <- convExpr arg
-            l' <- mkLamAbsScoped b $ convExpr body
-            pure $ PLC.Apply () l' a'
-        GHC.Let (GHC.Rec bs) body -> do
-            -- See note [Recursive lets]
-            -- TODO: we're technically using these twice, which is bad and maybe we need to duplicate
-            tys <- mapM convType (fmap (GHC.varType . fst) bs)
-            -- The pairs of types we'll need for fixN
-            asbs <- forM tys $ \case
-                PLC.TyFun () i o -> pure (i, o)
-                _ -> throwPlain $ ConversionError "Recursive values must be of function type. You may need to manually add unit arguments."
-
-            -- We need this so we can use the tuple of recursive functions in the end - it
-            -- expects an output type.
-            outTy <- convType (GHC.exprType body)
-
-            q <- safeFreshTyName "Q"
-            choose <- safeFreshName "choose"
-            let chooseTy = mkIterTyFun tys (PLC.TyVar () q)
-
-            -- \f1 ... fn -> choose b1 ... bn
-            bsLam <- mkIterLamAbsScoped (fmap fst bs) $ do
-                rhss <- mapM convExpr (fmap snd bs)
-                pure $ mkIterApp (PLC.Var() choose) rhss
-
-            -- abstract out Q and choose
-            let cLam = PLC.TyAbs () q (PLC.Type ()) $ PLC.LamAbs () choose chooseTy bsLam
-
-            -- fixN {A1 B1 ... An Bn}
-            instantiatedFix <- do
-                fixN <- liftQuote $ Function.getBuiltinFixN (length bs)
-                pure $ mkIterInst fixN $ foldMap (\(a, b) -> [a, b]) asbs
-
-            let fixed = PLC.Apply () instantiatedFix cLam
-
-            -- the consumer of the recursive functions
-            bodyLam <- mkIterLamAbsScoped (fmap fst bs) (convExpr body)
-            pure $ PLC.Apply () (PLC.TyInst () fixed outTy) bodyLam
-        GHC.Case scrutinee b t alts -> do
-            let scrutineeType = GHC.varType b
-
-            -- See Note [Case expressions and laziness]
-            isValueAlt <- mapM (\(_, vars, body) -> if null vars then PLC.isTermValue <$> convExpr body else pure True) alts
-            let lazyCase = not $ and isValueAlt
-
-            scrutinee' <- convExpr scrutinee
-
-            match <- getMatchInstantiated scrutineeType
-            let matched = PLC.Apply () match scrutinee'
-
-            -- See Note [Scott encoding of datatypes]
-            -- we're going to delay the body, so the scrutinee needs to be instantiated the delayed type
-            instantiated <- PLC.TyInst () matched <$> (convType t >>= maybeDelayType lazyCase)
-
-            tc <- case GHC.splitTyConApp_maybe scrutineeType of
-                Just (tc, _) -> pure tc
-                Nothing      -> throwPlain $ ConversionError "Scrutinee's type was not a type constructor application"
-            dcs <- getDataCons tc
-
-            branches <- forM dcs $ \dc -> case GHC.findAlt (GHC.DataAlt dc) alts of
-                Just alt -> convAlt lazyCase dc alt
-                Nothing  -> throwPlain $ ConversionError "No case matched and no default case"
-
-            let applied = mkIterApp instantiated branches
-            -- See Note [Case expressions and laziness]
-            maybeForce lazyCase applied
-        -- ignore annotation
-        GHC.Tick _ body -> convExpr body
-        GHC.Cast body coerce -> do
-            body' <- convExpr body
-            case splitNewtypeCoercion coerce of
-                Just (Unwrap outer inner) -> do
-                    match <- getMatchInstantiated outer
-                    -- unwrap by doing a "trivial match" - instantiate to the inner type and apply the identity
-                    inner' <- convType inner
-                    let instantiated = PLC.TyInst () (PLC.Apply () match body') inner'
-                    name <- safeFreshName "inner"
-                    let identity = PLC.LamAbs () name inner' (PLC.Var () name)
-                    pure $ PLC.Apply () instantiated identity
-                Just (Wrap _ outer) -> do
-                    constr <- head <$> getConstructorsInstantiated outer
-                    pure $ PLC.Apply () constr body'
-                _ -> throwSd UnsupportedError $ "Coercion" GHC.$+$ GHC.ppr coerce
-        GHC.Type _ -> throwPlain $ ConversionError "Cannot convert types directly, only as arguments to applications"
-        GHC.Coercion _ -> throwPlain $ ConversionError "Coercions should not be converted"
-
-convExprWithDefs :: Converting m => GHC.CoreExpr -> m PLCTerm
-convExprWithDefs e = do
-    converted <- convExpr e
-    ConvertingState{csTypeDefs=typeDs, csTermDefs=termDs} <- get
-    wrapWithDefs typeDs termDs converted
