--- core-to-plc/src/Language/Plutus/CoreToPLC/Primitives.hs
@@ -1,39 +1,0 @@
--- | Primitive names that correspond to Plutus Core primitives.
-module Language.Plutus.CoreToPLC.Primitives where
-
-import           Language.Plutus.CoreToPLC.Error
-
--- | An abstract data type representing bytestrings in Plutus Core.
-data ByteString
-
--- TODO: resizing primitives? better handling of sizes?
-
-concatenate :: ByteString -> ByteString -> ByteString
-concatenate = mustBeReplaced
-
-takeByteString :: Int -> ByteString -> ByteString
-takeByteString = mustBeReplaced
-
-dropByteString :: Int -> ByteString -> ByteString
-dropByteString = mustBeReplaced
-
-sha2_256 :: ByteString -> ByteString
-sha2_256 = mustBeReplaced
-
-sha3_256 :: ByteString -> ByteString
-sha3_256 = mustBeReplaced
-
-verifySignature :: ByteString -> ByteString -> ByteString -> Bool
-verifySignature = mustBeReplaced
-
-equalsByteString :: ByteString -> ByteString -> Bool
-equalsByteString = mustBeReplaced
-
-txhash :: ByteString
-txhash = mustBeReplaced
-
-blocknum :: Int
-blocknum = mustBeReplaced
-
-error :: () -> a
-error = mustBeReplaced
